pyproject.toml
[tool.poetry]
name = "rural-producers"
version = "0.1.0"
description = "Sistema de gerenciamento de produtores rurais"
authors = ["Jackson Cypriano <jacksoncypriano@hotmail.com>"]
packages = [{ include = "core" }]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
django = "^5.0"
validate-docbr = "^1.11.1"
djangorestframework = "^3.16.0"
python-dotenv = "^1.1.1"
psycopg2-binary = "^2.9.10"
gunicorn = "^21.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
model-bakery = "^1.20.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
